add auth_user to authenticate and get_user response
