+++ api/src/auth/mod.rs
@@ -16,7 +16,7 @@ pub use self::{
   users::{AuthBackend, AuthSession},
   web::{login_post_internal, logout_post_internal},
 };
-use crate::{sessions::MySessionManagerLayer, ApiError, ApiResult, MINIMUM_KARMA_TO_DOWNVOTE};
+use crate::{sessions::MySessionManagerLayer, ApiError, ApiResult};
 
 pub type MyAuthLayer = AuthManagerLayer<AuthBackend, PostgresStore, SignedCookie>;
 
@@ -52,43 +52,4 @@ impl AuthenticationExt for AuthSession {
   fn am_authenticated_and_not_banned(&self) -> bool {
     self.user.as_ref().map(|user| !user.0.banned).unwrap_or(false)
   }
-}
-
-/// Local authentication state, mirroring middleware on reference implementation
-///
-/// ref: https://github.com/thor314/zkhn/blob/main/rest-api/middlewares/index.js#L36
-#[derive(Debug, Clone, Default, Serialize, Deserialize, ToSchema)]
-#[schema(default = AuthLocal::default, example=AuthLocal::default)]
-#[serde(rename_all = "camelCase")]
-pub struct AuthLocal {
-  pub user_signed_in:   bool,
-  pub username:         Option<Username>,
-  pub karma:            Option<i32>,
-  pub contains_email:   Option<bool>,
-  pub show_dead:        bool,
-  pub show_downvote:    bool,
-  pub is_moderator:     Option<bool>,
-  // shadow_banned: removed
-  pub banned:           bool,
-  pub cookies_included: bool,
-}
-
-impl AuthLocal {
-  /// Create a new AuthLocal from a User
-  pub fn new(user: User) -> Self {
-    Self {
-      user_signed_in:   true,
-      username:         Some(user.username.clone()),
-      karma:            Some(user.karma),
-      contains_email:   Some(user.email.is_some()),
-      show_dead:        user.show_dead,
-      show_downvote:    user.karma >= MINIMUM_KARMA_TO_DOWNVOTE,
-      is_moderator:     Some(user.is_moderator),
-      banned:           user.banned,
-      cookies_included: true,
-    }
-  }
-
-  /// Create a new AuthLocal without authentication
-  pub fn new_unauthenticated(banned: bool) -> Self { Self { banned, ..Default::default() } }
 }
